feat(FE-api-layer): 스키마 domain, features구분에 따른 디렉토리 수정 및 적용
